fix(core): Avoid buffering on the queue
Using bufferTime(0) doesn't work because a buffer is created whenever the
previous one closes, *not* when a new emission comes along. This causes
an endless stream of emissions, which kills performance.

We now instead use the same approach as the Forms API and as in other
places by passing some emitEvent flag.

fixes #23