fix(core): Improve typing and support emptyOn for multi:true
This change improves the type safety by not misinterpreting an
array-valued parameter as a multi: true parameter. We also introduce a new
MultiQueryParam type as a shortcut.

As a side effect, we move the evaluation of (de-)serializing entirely to
the service and can thus now support emptyOn for multi: true parameters.

fixes #92
fixes #27

Signed-off-by: Ingo Bürk <ingo.buerk@tngtech.com>