chore: remove support for schematics
We rely on schematics-utilities for this which has caused problems time and time again.
The latest problem is that this forces us to install schematics-utilities as a non-dev
dependency for library users, which e.g. ships an outdated version of kind-of popping up
in security scans (among other, similar issues).

We could solve this by adding schematic support in a separate @ngqp/install kind of package
which we can always do if the need arises. However, as it stands, installing ngqp is as
trivial as it gets anyway and the original intention of possibly requiring further configuration
has never turned out to be true.

Signed-off-by: Ingo Bürk <ingo.buerk@tngtech.com>